requirements.txt
click==7.1.2
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
python-dotenv==0.13.0
# tensorflow-gpu==1.15 # GPU Version of TensorFlow
# tensorflow==1.15.2   # CPU Version of TensorFlow
# tensorflow==2.2.0  # 2
# tensorflow_hub==0.7

# pip install torch==1.5.1+cpu torchvision==0.6.1+cpu -f https://download.pytorch.org/whl/torch_stable.html

sentencepiece==0.1.91
Flask==1.1.2
Flask-RESTful==0.3.8
transformers==3.0.1